Pace to float

+++ core/src/main/java/com/karyna/core/domain/run/Run.kt
@@ -14,6 +14,6 @@ data class Run(
     val coordinates: List<LatLng>,
     val durationS: Long,
     val distanceMeters: Int,
-    val paceMetersInS: Int,
+    val paceMetersInS: Float,
     val calories: Int?
 )
